+++ gradle.properties
@@ -1,3 +1,3 @@
 org.gradle.daemon=true
 org.gradle.jvmargs=-Xms256m -Xmx1024m
-archash=b38d5e9c24d921f87f33bc48fcb65e245c1b1d0c
+archash=190918590e8401b1686ecb9167e3c2a9e77eafaa
